source/requirements.txt
Adafruit_IO==2.7.0
pyserial==3.5
